Separate Enemy into its own file
